+++ src/test/sample/test.lua
@@ -137,3 +137,6 @@ local const_v = -16 + 1 << 32 + 8 >> "32" + 2 * 4 - 5 / 2 + 8 % 2
 
 -- test luaparse v0.3.0 encoding mode
 local str = "能不能正常解析中文utf8字符"
+
+-- do NOT jump to file battle_conf
+local BattleConfErr = require "attle_conf"